+ main: enable soft bounces for sometimes-disconnected operation
this never bounces mails "for real" but instead will keep retrying
indefinitely.  however at the same time we send mail to the local
postmaster indicating the problem.  this should work for
sometimes-disconnected scenario such as that on a laptop.  this is
subject to change later...

this was just the setting that was being used in 2016, last time the
config was changed, we're refreshing it now