- crypt: add my rsa key, this was an old repo that uses ed25519 key
we use an rsa4k key these days, not the ed25519, which we let expire.

rsa is more future-proof by way of quantum resistance than ed25519,
although supposedly we need a 15360 bit key to be post-quantum
safe-equivalent to post-quantum aes256 symmetric key, which is
equivalent to aes128 today in pre-quantum terms).  it seems that qubits
needed to crack a public RSA key of length N are 2N+3 by some published
algorithms, so we'll need a 8195-qubit [ideal, error-corrected] computer
to factor rsa4k, and it will have to stay coherent for some reasonable
amount of time (?), these are things that won't happen for a while.  we
can trade memory space for coherence time, but the memory requirements
become infeasible quickly, so supposedly 8k RSA keys will likely be good
for 50 years of post-quantum information security and 16k RSA keys for
100 years (see paper from kai-li and qing-yu cai)

since in any case, quantum is not an attack vector to be concerned about
just yet before the quantum-weak edd25519 is a problem and due to above
difficulties even once it's around, we have time to rekey before 4k is a
problem.

(rekey is mentioned because with git-crypt, old keys will continue to
work on old commits forever, even if there was a way to delete them,
which currently doesn't exist, see issue #47)